moved setOperations to utils
